Optimize notification service
